# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "AninfPush API Documentation",
        "description": "API microservice for managing multi-channel messaging (Email, SMS, WhatsApp) with Keycloak authentication",
        "contact": {
            "email": "support@aninfpush.com"
        },
        "license": {
            "name": "Proprietary",
            "url": "https://aninfpush.com/license"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/",
            "description": "API Server"
        }
    ],
    "paths": {
...
        "/api/v1/businesses/{businessId}/smtp-settings": {
            "get": {
                "tags": [
                    "SMTP Settings"
                ],
                "summary": "Get SMTP settings for a business",
                "description": "Returns all SMTP settings for a specific business",
                "operationId": "86f7d5b1f794bf1ac816cee0841a9d46",
                "parameters": [
                    {
                        "name": "businessId",
                        "in": "path",
                        "description": "Business ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/SmtpSetting"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "SMTP Settings"
                ],
                "summary": "Create a new SMTP setting",
                "description": "Creates a new SMTP setting for a business. Can reuse existing base configuration by providing base_smtp_id.",
                "operationId": "2cb1305c5da1ab70d66b8976fd202ecb",
                "parameters": [
                    {
                        "name": "businessId",
                        "in": "path",
                        "description": "Business ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "from_email",
                                    "from_name"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Primary SMTP"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "Main email configuration"
                                    },
                                    "base_smtp_id": {
                                        "description": "ID of existing SMTP to reuse base config",
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "host": {
                                        "type": "string",
                                        "example": "smtp.gmail.com"
                                    },
                                    "port": {
                                        "type": "integer",
                                        "example": 587
                                    },
                                    "encryption": {
                                        "type": "string",
                                        "enum": [
                                            "tls",
                                            "ssl",
                                            "none"
                                        ],
                                        "example": "tls"
                                    },
                                    "username": {
                                        "type": "string",
                                        "example": "user@example.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "example": "password123"
                                    },
                                    "from_email": {
                                        "type": "string",
                                        "example": "sender@example.com"
                                    },
                                    "from_name": {
                                        "type": "string",
                                        "example": "My Company"
                                    },
                                    "reply_to_email": {
                                        "type": "string",
                                        "example": "reply@example.com"
                                    },
                                    "reply_to_name": {
                                        "type": "string",
                                        "example": "Support"
                                    },
                                    "timeout": {
                                        "type": "integer",
                                        "example": 30
                                    },
                                    "verify_peer": {
                                        "type": "boolean",
                                        "example": true
                                    },
                                    "is_active": {
                                        "type": "boolean",
                                        "example": true
                                    },
                                    "is_default": {
                                        "type": "boolean",
                                        "example": false
                                    },
                                    "second_limit": {
                                        "type": "integer",
                                        "example": 10
                                    },
                                    "hourly_limit": {
                                        "type": "integer",
                                        "example": 1000
                                    },
                                    "daily_limit": {
                                        "type": "integer",
                                        "example": 10000
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "SMTP setting created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "SMTP setting created successfully"
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/SmtpSetting"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
        "/api/v1/smtp-settings/base": {
            "get": {
                "tags": [
                    "SMTP Settings"
                ],
                "summary": "Get available SMTP base configurations",
                "description": "Returns unique SMTP base configurations (host, port, encryption, username) that can be reused",
                "operationId": "504b24cc39ee73da914ee516fcd56e52",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer"
                                                    },
                                                    "host": {
                                                        "type": "string"
                                                    },
                                                    "port": {
                                                        "type": "integer"
                                                    },
                                                    "encryption": {
                                                        "type": "string"
                                                    },
                                                    "username": {
                                                        "type": "string"
                                                    },
                                                    "name": {
                                                        "type": "string"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
...
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Enter Keycloak JWT token",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}